pyproject.toml
[project]
name = "deep-agents-webapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.12",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.1.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
